Separate out CosmosDb into its own module
